VOC-63459: Aligned toolbar usage architecture for all samples
