+++ CustomInSessionInvite/src/main/java/com/verint/xm/demo/custominsessioninvite/BaseActivity.java
@@ -1,0 +1,29 @@
+package com.verint.xm.demo.custominsessioninvite;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.graphics.Insets;
+import androidx.core.view.ViewCompat;
+import androidx.core.view.WindowInsetsCompat;
+
+public abstract class BaseActivity extends AppCompatActivity {
+
+    protected void setupToolbar(boolean showBackButton) {
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        if (showBackButton) {
+            getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        }
+        ViewCompat.setOnApplyWindowInsetsListener(toolbar, (v, windowInsets) -> {
+            Insets statusBar = windowInsets.getInsets(WindowInsetsCompat.Type.statusBars());
+            v.setPadding(v.getPaddingLeft(), statusBar.top, v.getPaddingRight(), v.getPaddingBottom());
+            return windowInsets;
+        });
+    }
+
+    @Override
+    public boolean onSupportNavigateUp() {
+        onBackPressed();
+        return true;
+    }
+}
