+++ KotlinSample/src/main/java/com/verint/xm/kotlinsample/MainActivity.kt
@@ -2,30 +2,16 @@ package com.verint.xm.kotlinsample
 
 import android.os.Bundle
 import android.view.View
-import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.widget.Toolbar
-import androidx.core.graphics.Insets
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import com.verint.xm.sdk.Core
 import com.verint.xm.sdk.SurveyManagement
 
-class MainActivity : AppCompatActivity() {
+class MainActivity : BaseActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.main)
 
-        val toolbar = findViewById<Toolbar>(R.id.toolbar)
-        // Toolbar is used as the ActionBar to show the activity title
-        setSupportActionBar(toolbar)
-        // Expand the Toolbar top padding to equal the status bar height, so the Toolbar background fills behind the status bar.
-        // Required by API 35+.
-        ViewCompat.setOnApplyWindowInsetsListener(toolbar) { v, windowInsets ->
-            val statusBar = windowInsets.getInsets(WindowInsetsCompat.Type.statusBars())
-            v.setPadding(v.paddingLeft, statusBar.top, v.paddingRight, v.paddingBottom)
-            windowInsets
-        }
+        setupToolbar()
     }
